+++ test/Geo.ArcGIS.Tests/Services/ArcGISTokenRetrevialShould.cs
@@ -7,7 +7,6 @@ namespace Geo.ArcGIS.Tests.Services
 {
     using System;
     using System.Collections.Generic;
-    using System.Linq;
     using System.Net;
     using System.Net.Http;
     using System.Threading;